pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["scclone2dr*"]

[project]
name = "scClone2DR"
version = "0.1.0"
description = "Clone-level multi-modal prediction of tumour drug response"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Quentin Duchemin", email = "qduchemin9@gmail.com"}
]
keywords = ["single-cell", "drug-response", "bioinformatics", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.20.0",
    "torch>=1.10.0",
    "pyro-ppl>=1.8.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "h5py>=3.0.0",
    "tqdm>=4.60.0",
    "scikit-learn>=0.24.0",
    "scikit-fda>=0.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12",
    "black>=21.0",
    "flake8>=3.9",
    "isort>=5.9",
    "mypy>=0.910",
]

notebook = [
    "ipykernel",
    "jupyter",
    "nbformat>=5.0.0"
]

[project.urls]
Homepage = "https://github.com/cbg-ethz/scClone2DR"
Documentation = "https://scClone2DR.readthedocs.io"
Repository = "https://github.com/cbg-ethz/scClone2DR"
"Bug Tracker" = "https://github.com/cbg-ethz/scClone2DR/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=scclone2dr --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
